+++ spec/ncs_navigator/warehouse/data_mapper_spec.rb
@@ -40,4 +40,31 @@ module NcsNavigator::Warehouse::DataMapper
 
     it_behaves_like 'a string-based NCS type'
   end
+
+  describe NcsInteger, :modifies_warehouse_state do
+    let(:model) {
+      class NcsintegerHost
+        include DataMapper::Resource
+        property :code, NcsInteger
+      end
+    }
+
+    let(:prop) { model.properties[:code] }
+
+    it 'can be set from a canonical string' do
+      prop.send(:typecast_to_primitive, '46').should == 46
+    end
+
+    it 'can be set from a Fixnum' do
+      prop.send(:typecast_to_primitive, 50).should == 50
+    end
+
+    it 'converts to an integer even if there are leading zeros' do
+      prop.send(:typecast_to_primitive, '037').should == 37
+    end
+
+    it 'does not convert to an integer if it is a non-numeric string' do
+      prop.send(:typecast_to_primitive, 'abc123').should == 'abc123'
+    end
+  end
 end